(feat-2-tela-para-entrar-numa-sala-por-código): Dividindo tela para ver a sala por tipo de usuário
