+++ src/interfaces/user.ts
@@ -1,0 +1,5 @@
+
+export interface IUser {
+  name: string;
+  roomCode: string;
+};
