+++ src/services/auth.ts
@@ -1,0 +1,4 @@
+
+export async function verifyIsHost(): Promise<boolean> {
+  return false;
+}
